+++ src/lib.rs
@@ -151,6 +151,7 @@ fn set_x52pro_inputs_from_ship_statues(
             Attribute::Hardpoints => vec![Control::Hardpoints],
             Attribute::HeatSink => vec![Control::HeatSink],
             Attribute::LandingGear => vec![Control::LandingGear],
+            Attribute::NightVision => vec![Control::NightVision],
             Attribute::SilentRunning => vec![Control::SilentRunning],
             Attribute::Throttle => vec![Control::Throttle],
         }